requirements.txt
google-adk
google-genai
google-cloud 
google-cloud-bigquery
google-cloud-storage
pytest
python-dotenv>=1.0.1